Use a serial queue for messaging
The polling for synchronous messaging and the push operations, are now scheduled in a serial queue. A serial queue give us a lock free code but potentially less performances, however as libwebsockets doesn't give us concurrent I/O the performance penalty will only be visible when a lot of processing is done in the handleRequestBlock.